Integrate TMP36 temperature sensor with Arduino
This commit adds support for the TMP36 sensor to measure temperature, enhancing the weather station's data set. A new pin definition was incorporated for TMP36 and an input initialization was included in the setup. Temperature reading from sensor is implemented in the execution flow of the main loop for constant updates. Furthermore, a new condition in the LCD print and an update in the menu functionalities were made to display the acquired temperature readings in Celsius. This enhancement provides more extensive meteorological data set.